Fix cyclic test dependency
