+++ lib/multiplexer/multiplexer_test.go
@@ -50,6 +50,7 @@ import (
 	"github.com/gravitational/teleport/lib/sshutils"
 	"github.com/gravitational/teleport/lib/tlsca"
 	"github.com/gravitational/teleport/lib/utils"
+	tlsutils "github.com/gravitational/teleport/lib/utils/tls"
 )
 
 func TestMain(m *testing.M) {
@@ -60,7 +61,7 @@ func TestMain(m *testing.M) {
 // TestMux tests multiplexing protocols
 // using the same listener.
 func TestMux(t *testing.T) {
-	_, signer, err := utils.CreateCertificate("foo", ssh.HostCert)
+	_, signer, err := tlsutils.CreateCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
 	// TestMux tests basic use case of multiplexing TLS
