+++ lib/srv/regular/sshserver_test.go
@@ -70,6 +70,7 @@ import (
 	"github.com/gravitational/teleport/lib/sshutils"
 	"github.com/gravitational/teleport/lib/sshutils/x11"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/tls"
 )
 
 // teleportTestUser is additional user used for tests
@@ -1204,7 +1205,7 @@ func TestKeyAlgorithms(t *testing.T) {
 	t.Parallel()
 	f := newFixture(t)
 
-	_, ellipticSigner, err := utils.CreateEllipticCertificate("foo", ssh.UserCert)
+	_, ellipticSigner, err := tls.CreateEllipticCertificate("foo", ssh.UserCert)
 	require.NoError(t, err)
 
 	sshConfig := &ssh.ClientConfig{
