+++ lib/sshutils/server_test.go
@@ -29,6 +29,7 @@ import (
 	"golang.org/x/crypto/ssh"
 
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/utils/tls"
 )
 
 func TestMain(m *testing.M) {
@@ -39,7 +40,7 @@ func TestMain(m *testing.M) {
 func TestStartStop(t *testing.T) {
 	t.Parallel()
 
-	_, signer, err := utils.CreateCertificate("foo", ssh.HostCert)
+	_, signer, err := tls.CreateCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
 	called := false
@@ -91,7 +92,7 @@ func TestStartStop(t *testing.T) {
 func TestShutdown(t *testing.T) {
 	t.Parallel()
 
-	_, signer, err := utils.CreateCertificate("foo", ssh.HostCert)
+	_, signer, err := tls.CreateCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
 	closeContext, cancel := context.WithCancel(context.TODO())
@@ -147,7 +148,7 @@ func TestShutdown(t *testing.T) {
 func TestConfigureCiphers(t *testing.T) {
 	t.Parallel()
 
-	_, signer, err := utils.CreateCertificate("foo", ssh.HostCert)
+	_, signer, err := tls.CreateCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
 	fn := NewChanHandlerFunc(func(_ context.Context, _ *ConnectionContext, nch ssh.NewChannel) {
@@ -196,10 +197,10 @@ func TestConfigureCiphers(t *testing.T) {
 func TestHostSignerFIPS(t *testing.T) {
 	t.Parallel()
 
-	_, signer, err := utils.CreateCertificate("foo", ssh.HostCert)
+	_, signer, err := tls.CreateCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
-	_, ellipticSigner, err := utils.CreateEllipticCertificate("foo", ssh.HostCert)
+	_, ellipticSigner, err := tls.CreateEllipticCertificate("foo", ssh.HostCert)
 	require.NoError(t, err)
 
 	fn := NewChanHandlerFunc(func(_ context.Context, _ *ConnectionContext, nch ssh.NewChannel) {
